+++ src/protocol/index.ts
@@ -1,2 +1,3 @@
 export { AddressComponentRangeError, MEMORY_REGIONS, MultiSlot, PresetSlot } from "./address";
 export * from "./cc";
+export * from "./enums";